Changed type from short to int to allow for larger quantities

+++ Skyblock/Analyzer/com/rithsagea/skyblock/api/datatypes/Datapoint.java
@@ -6,9 +6,9 @@ public class Datapoint {
 	
 	public Timestamp time = null;
 	public double value = 0;
-	public short amount = 0;
+	public int amount = 0;
 	
-	public Datapoint(Timestamp time, double value, short amount) {
+	public Datapoint(Timestamp time, double value, int amount) {
 		this.time = time;
 		this.value = value;
 		this.amount = amount;
